+++ flask_security/core.py
@@ -1634,6 +1634,7 @@ class Security:
 
         # If they don't want ALL mechanisms protected, then they must
         # set WTF_CSRF_CHECK_DEFAULT=False so that our decorators get control.
+        # And our decorators use csrf.protect()
         if cv("CSRF_PROTECT_MECHANISMS", app=app) != AUTHN_MECHANISMS:
             if not csrf:
                 # This isn't good.