+++ CHANGES.rst
@@ -13,6 +13,7 @@ Fixes
 - (:issue:`950`) Regression - some templates no longer getting correct config (thanks pete7863).
 - (:issue:`954`) CSRF not properly ignored for application forms using SECURITY_CSRF_PROTECT_MECHANISMS.
 - (:pr:`957`) Improve jp translations (e-goto)
+- (:issue:`959`) Regression - datetime_factory should still be an attribute (thanks TimotheeJeannin)
 
 Version 5.4.2
 -------------